libft

+++ ft_strdup.c
@@ -6,28 +6,28 @@
 /*   By: hkrifa <hkrifa@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/03/24 11:22:13 by hkrifa            #+#    #+#             */
-/*   Updated: 2021/03/24 12:03:29 by hkrifa           ###   ########.fr       */
+/*   Updated: 2021/03/28 12:53:30 by haroun           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-char    *ft_strdup(const char *src)
+char	*ft_strdup(const char *src)
 {
-    int i;
-    int len;
-    char *dest;
+	size_t	i;
+	size_t	len;
+	char	*dest;
 
-    len = ft_strlen((char *)src);
-    dest = malloc(sizeof(char) * len + 1);
-    if(dest == NULL)
-        return (0);
-    i = 0; 
-    while(i <= len)
-    {
-        dest[i] = src[i];
-        i++;
-    }
-    dest[i] = '\0';
-    return(dest);
+	len = ft_strlen((char *)src);
+	dest = malloc(sizeof(char) * len + 1);
+	if (dest == NULL)
+		return (0);
+	i = 0;
+	while (i <= len)
+	{
+		dest[i] = src[i];
+		i++;
+	}
+	dest[i] = '\0';
+	return (dest);
 }
